Calculate the forecasts for a VAR out to a specified horizon Conditions on information at time "t"

Calculate the forecasts for a VAR out to a specified horizon Conditions on information at time "t"

Usage

forecast(var, horizon = 1)

Arguments

var

vars::VAR or svars object

horizon

number of steps out to calculate forecast

Value

forecast

Examples

x <- svars::USA
v <- vars::VAR(x, p = 2)
forecast(v, 10)
#>                  x       pi         i
#>   [1,]          NA       NA        NA
#>   [2,]          NA       NA        NA
#>   [3,]          NA       NA        NA
#>   [4,]          NA       NA        NA
#>   [5,]          NA       NA        NA
#>   [6,]          NA       NA        NA
#>   [7,]          NA       NA        NA
#>   [8,]          NA       NA        NA
#>   [9,]          NA       NA        NA
#>  [10,]          NA       NA        NA
#>  [11,]          NA       NA        NA
#>  [12,]  1.41039689 4.038803  5.510138
#>  [13,]  1.62717789 4.182735  5.634644
#>  [14,]  1.60712751 4.925727  6.458551
#>  [15,]  1.65911647 5.342172  6.945378
#>  [16,]  1.42745525 5.336419  6.881628
#>  [17,]  0.96680536 5.495716  7.292802
#>  [18,]  0.92760553 5.247775  7.065628
#>  [19,]  1.58478954 4.483222  5.933315
#>  [20,]  1.80326189 4.261544  5.455047
#>  [21,]  1.29354490 5.040519  6.525950
#>  [22,]  0.85668056 5.397753  7.099772
#>  [23,]  0.67077699 5.700933  7.665663
#>  [24,]  0.42902057 5.683484  7.849866
#>  [25,]  0.59313860 5.326056  7.291113
#>  [26,]  0.16969638 5.806381  7.911956
#>  [27,]  0.10476651 5.551758  7.865507
#>  [28,] -0.68038564 5.459146  8.151410
#>  [29,] -1.12072349 5.452480  8.356867
#>  [30,] -1.21444119 4.869074  7.665033
#>  [31,] -1.37197315 4.721323  7.532785
#>  [32,] -1.32442924 4.623351  7.362805
#>  [33,] -0.47398893 3.846407  6.177501
#>  [34,] -0.55234872 3.993667  6.064405
#>  [35,] -0.19995310 5.227535  7.355187
#>  [36,] -0.36273219 5.071207  7.242681
#>  [37,] -0.39071880 4.407092  6.674265
#>  [38,]  0.12983237 3.756194  5.632443
#>  [39,]  0.10509193 5.180427  7.102875
#>  [40,]  0.53564782 4.549533  6.512316
#>  [41,]  0.57271097 4.544207  6.388941
#>  [42,]  0.16041919 5.367710  7.501263
#>  [43,] -0.13586046 5.691310  8.199136
#>  [44,] -0.68297296 6.060500  8.789527
#>  [45,] -2.31333899 6.223269  9.727404
#>  [46,] -2.24565683 6.241907  9.678225
#>  [47,] -2.17643401 5.723652  8.889058
#>  [48,] -3.58013404 6.327063 10.384812
#>  [49,] -4.70887935 6.624542 11.039837
#>  [50,] -4.10412691 6.559837 10.408433
#>  [51,] -2.83580143 5.294286  8.250887
#>  [52,] -1.93807923 4.142156  6.873425
#>  [53,] -2.16483491 4.558000  7.576245
#>  [54,] -1.69475306 4.563109  7.272676
#>  [55,] -0.72185353 4.021510  6.393445
#>  [56,] -0.56664920 3.741127  5.981401
#>  [57,] -0.83658446 4.039333  6.362036
#>  [58,] -1.19189326 4.902403  7.350974
#>  [59,] -1.11936304 5.004946  7.463315
#>  [60,] -0.94147464 4.922048  7.454249
#>  [61,] -0.76824210 4.716757  7.234260
#>  [62,] -1.78873520 5.778843  8.614740
#>  [63,] -1.72689812 4.883211  7.718941
#>  [64,] -1.66621260 5.879525  9.074356
#>  [65,] -1.72447487 5.662974  8.738694
#>  [66,] -2.53097921 6.257872  9.843812
#>  [67,] -2.58069755 5.568256  9.076407
#>  [68,] -3.17564546 6.182945  9.908729
#>  [69,] -3.51336005 5.967488  9.975177
#>  [70,] -4.36682243 5.340952  9.915070
#>  [71,] -4.90903242 4.999306  9.866281
#>  [72,] -4.36437310 4.372915  8.447218
#>  [73,] -3.77908499 4.772146  8.507824
#>  [74,] -6.65653844 6.156640 12.064277
#>  [75,] -6.24084233 5.623895 11.282949
#>  [76,] -6.10746424 4.017803  9.458995
#>  [77,] -5.71617377 3.643812  8.999749
#>  [78,] -4.36475551 3.314654  7.435249
#>  [79,] -4.90799439 2.404755  6.853656
#>  [80,] -4.86876980 1.793585  6.389280
#>  [81,] -3.74565514 1.802870  5.425482
#>  [82,] -3.23165004 1.667213  5.032864
#>  [83,] -2.69096316 1.431484  4.667040
#>  [84,] -2.34423276 1.588801  4.831137
#>  [85,] -2.59233937 2.383930  5.806554
#>  [86,] -2.15274438 2.332759  5.594221
#>  [87,] -2.17753748 3.020955  6.345420
#>  [88,] -2.21981321 3.021105  6.463881
#>  [89,] -2.30618594 2.800651  6.278942
#>  [90,] -1.27817405 2.511250  5.277268
#>  [91,] -1.41490070 3.495301  6.327729
#>  [92,] -0.93477902 3.002188  5.602179
#>  [93,] -0.66050895 2.970204  5.530839
#>  [94,] -0.68714217 2.907957  5.436390
#>  [95,] -0.49330848 2.772890  5.198310
#>  [96,] -0.09004280 2.541715  4.658095
#>  [97,]  0.15763835 2.675382  4.688745
#>  [98,] -0.05903980 2.897393  4.988784
#>  [99,] -0.26661149 3.188846  5.375084
#> [100,] -0.42103100 3.258879  5.604802
#> [101,] -0.46454795 3.292991  5.648546
#> [102,] -0.47469053 3.682281  6.127782
#> [103,] -0.32827241 3.581762  5.846505
#> [104,] -0.62794476 4.015775  6.545700
#> [105,] -1.14500185 4.286454  7.047425
#> [106,] -0.99720061 3.846806  6.645781
#> [107,] -1.40421497 4.124616  7.159773
#> [108,] -1.48916059 4.040220  7.096679
#> [109,] -0.93718200 3.512003  6.248785
#> [110,] -0.72958305 3.268862  5.827344
#> [111,] -0.93740604 4.047247  6.751221
#> [112,] -1.04805060 4.054672  6.760571
#> [113,] -0.99189532 3.617802  6.241431
#> [114,] -0.92159493 2.936789  5.370381
#> [115,] -0.84657123 3.017075  5.285698
#> [116,] -0.59563432 2.841957  5.094682
#> [117,] -0.49745875 2.792659  4.966229
#> [118,] -0.08883037 2.494185  4.403752
#> [119,]  0.37918772 2.493591  4.216507
#> [120,]  0.38904637 2.964482  4.689132
#> [121,]  0.70782398 2.916408  4.461431
#> [122,]  0.71313744 3.296187  4.841794
#> [123,]  0.68054818 3.091928  4.583133
#> [124,]  0.64347740 3.080017  4.608162
#> [125,]  0.56952176 3.029433  4.582058
#> [126,]  0.72696403 3.066679  4.625438
#> [127,]  0.76903622 3.042778  4.588960
#> [128,]  0.59103845 3.179810  4.920335
#> [129,]  0.40482207 3.109745  4.898407
#> [130,]  0.18291680 3.178382  5.168995
#> [131,] -0.07409866 3.009599  5.080154
#> [132,] -0.09939938 2.632070  4.701006
#> [133,] -0.00425289 2.642955  4.679000
#> [134,] -0.01080821 2.695332  4.720617
#> [135,]  0.06363894 2.758549  4.707904
#> [136,]  0.31658141 2.842877  4.797288
#> [137,]  0.50320656 2.661045  4.485506
#> [138,]  0.40931309 3.023239  4.897502
#> [139,]  0.25504229 3.281661  5.202495
#> [140,]  0.49157315 2.983980  4.909332
#> [141,]  0.65296928 3.053382  4.890853
#> [142,]  0.68229310 3.028308  4.811884
#> [143,]  0.83654571 2.806807  4.545969
#> [144,]  0.88754599 2.835447  4.548472
#> [145,]  0.78993260 3.227268  5.025322
#> [146,]  1.18527926 3.416657  5.024116
#> [147,]  1.21357635 3.609896  5.162008
#> [148,]  1.22557311 3.549260  5.096555
#> [149,]  1.13131351 3.673200  5.339676
#> [150,]  1.11052724 4.011103  5.758143
#> [151,]  0.69033841 4.296230  6.126798
#> [152,]  0.61076425 4.356361  6.426773
#> [153,]  0.51531447 4.100012  6.054298
#> [154,]  0.50603879 3.847065  5.812156
#> [155,]  0.62780284 3.653202  5.372263
#> [156,]  0.86644079 3.815736  5.384136
#> [157,]  1.22793736 3.120135  4.392192
#> [158,]  1.69807084 2.885358  3.858022
#> [159,]  1.74189385 3.090282  4.108755
#> [160,]  1.62238896 3.157408  4.204741
#> [161,]  1.48505775 3.193187  4.284314
#> [162,]  1.37936061 3.176443  4.227517
#> [163,]  1.25756643 3.296335  4.420611
#> [164,]  1.48453444 2.930587  4.009208
#> [165,]  1.60365055 3.457508  4.543154
#> [166,]  1.69667530 3.605349  4.635515
#> [167,]  1.44498144 4.118379  5.215692
#> [168,]  1.36103328 4.268154  5.428243
#> [169,]  1.40364563 4.032583  5.236962
#> [170,]  1.26854477 4.080741  5.385562
#> [171,]  0.99508295 4.458709  5.952264
#> [172,]  0.98193616 4.127458  5.618516
#> [173,]  0.68039999 4.436247  6.118397
#> [174,]  0.57472892 4.195587  5.943269
#> [175,]  0.55449855 4.202967  6.057362
#> [176,]  0.38508936 4.105239  5.984115
#> [177,]  0.27252278 3.747227  5.662149
#> [178,]  0.59183536 3.195138  5.018153
#> [179,]  0.01442462 4.074719  6.067328
#> [180,]  0.20824944 3.591892  5.572568
#> [181,]  0.65668767 3.048308  4.803605
#> [182,]  0.83825938 3.074218  4.663823
#> [183,]  1.03112823 3.040037  4.311489
#> [184,]  1.31121898 3.149579  4.325017
#> [185,]  1.03940751 3.264132  4.471843