Impulse Responses.

Alias for the function from svars::irf.svars, so that the shock names are correct.

Usage

# S3 method for fevdvar
irf(
  x,
  impulse = NULL,
  response = NULL,
  n.ahead = 10,
  ortho = TRUE,
  cumulative = FALSE,
  boot = TRUE,
  ci = 0.95,
  runs = 100,
  seed = NULL,
  ...
)

# S3 method for bvartools
irf(
  x,
  impulse = NULL,
  response = NULL,
  n.ahead = 10,
  ortho = TRUE,
  cumulative = FALSE,
  boot = TRUE,
  ci = 0.95,
  runs = 100,
  seed = NULL,
  ...
)

Arguments

x

SVAR object of class "fevdvar".

impulse

A character vector of the impulses, default is all variables.

response

A character vector of the responses, default is all variables.

n.ahead

Integer specifying the steps.

ortho

Not used. Here to match generic vars::irf.

cumulative

Not used. Here to match generic vars::irf.

boot

Not used. Here to match generic vars::irf.

ci

Not used. Here to match generic vars::irf.

runs

Not used. Here to match generic vars::irf.

seed

Not used. Here to match generic vars::irf.

...

Currently not used.

Value

Matrix of forecast error variance decomposition in frequency domain Indexed: frequencies, variables, shocks

Examples

x <- svars::USA
v <- vars::VAR(x, p = 2)
mvar <- id_fevdtd(v, "pi", 4:10)
vars::irf(mvar)
#> 
#> Impulse response coefficients
#>    V1 epsilon[ Main ]%->%x epsilon[ Orth_2 ]%->%x epsilon[ Orth_3 ]%->%x
#> 1   0           0.06778726             0.71291235             0.08827537
#> 2   1           0.09962456             0.80865766             0.04477040
#> 3   2           0.06889548             0.74259796             0.09282645
#> 4   3           0.02815543             0.61125699             0.16134435
#> 5   4          -0.02473531             0.47068765             0.22229865
#> 6   5          -0.08041218             0.34340869             0.26958530
#> 7   6          -0.13571717             0.23639103             0.30106211
#> 8   7          -0.18797320             0.14994059             0.31833927
#> 9   8          -0.23567840             0.08167321             0.32377205
#> 10  9          -0.27791359             0.02851285             0.31986174
#>    epsilon[ Main ]%->%pi epsilon[ Orth_2 ]%->%pi epsilon[ Orth_3 ]%->%pi
#> 1              1.0641672             -0.18342601             0.111448857
#> 2              0.7153101             -0.07757648            -0.092187863
#> 3              0.7588303              0.00572451            -0.031024691
#> 4              0.7031149              0.05600512            -0.042662458
#> 5              0.6692832              0.09303648            -0.025868304
#> 6              0.6283811              0.11558081            -0.010405058
#> 7              0.5871757              0.12832824             0.008532501
#> 8              0.5444260              0.13353568             0.027944614
#> 9              0.5008284              0.13330069             0.047148780
#> 10             0.4567058              0.12913928             0.065326409
#>    epsilon[ Main ]%->%i epsilon[ Orth_2 ]%->%i epsilon[ Orth_3 ]%->%i
#> 1             0.2729854              0.3480848             -0.8051429
#> 2             0.3715669              0.6264262             -0.7947887
#> 3             0.4911181              0.6596073             -0.7260503
#> 4             0.5544104              0.6126489             -0.6355488
#> 5             0.5988322              0.5402418             -0.5361566
#> 6             0.6232267              0.4669393             -0.4443058
#> 7             0.6336541              0.4024302             -0.3620567
#> 8             0.6327544              0.3488524             -0.2902260
#> 9             0.6229053              0.3054276             -0.2278422
#> 10            0.6058520              0.2703395             -0.1736920