Find the IRF in FD for a VAR

Find the IRF in FD for a VAR

Usage

irffd_generic(var, freqs = c(0, 2 * pi), grid_size = 1000)

Arguments

var

A VAR. Currently supports either a 'svars' or 'fevdvar' object.

freqs

vector of length 2 of min and max frequencies (0:2pi)

grid_size

how fine the grid to approximate the frequency domain

Value

list with vector of frequencies and matrix of irf values